llm_backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
pydantic-settings
sentence-transformers
pypdf
langchain-text-splitters
faiss-cpu
langchain-community
langchain-huggingface
